Fix -Wall warnings.
git-svn-id: https://tinymux.googlecode.com/svn/branches/dev_brazil@231 d1b986fa-651c-0410-a323-35a8662cf44d

+++ mux/src/alloc.h
@@ -49,7 +49,7 @@ extern void pool_reset(void);
 
 #define safe_copy_chr(src, buff, bufp, nSizeOfBuffer) \
 { \
-    if ((*bufp - buff) < nSizeOfBuffer) \
+    if ((size_t)(*bufp - buff) < nSizeOfBuffer) \
     { \
         **bufp = src; \
         (*bufp)++; \
